Add : Création et suppresion d'avis

+++ app/Http/Controllers/ReviewController.php
@@ -4,46 +4,55 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Review;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class ReviewController extends Controller
 {
-    // Méthode pour récupérer tous les avis
-    public function index()
-    {
-        $reviews = Review::all();
-        return response()->json($reviews);
-    }
-
-    // Méthode pour créer un nouvel avis
     public function store(Request $request)
     {
-        $review = Review::create($request->all());
+        $request->validate([
+            'message' => 'required|string',
+            'reviewed_user_id' => 'required|exists:users,id',
+        ]);
+    
+        $user = Auth::user();
+        $reviewedUserId = $request->input('reviewed_user_id');
+        $message = $request->input('message');
+        
+        $review = new Review([
+            'message' => $message,
+            'user_id' => $user->id,
+            'reviewed_user_id' => $reviewedUserId,
+        ]);
+    
+        $review->save();
+    
         return response()->json($review, 201);
     }
 
-    // Méthode pour récupérer un avis par son ID
-    public function show($id)
+    public function getUserReviews($userId)
     {
-        $review = Review::findOrFail($id);
-        return response()->json($review);
-    }
+        $user = User::find($userId);
 
-    // Méthode pour mettre à jour un avis
-    public function update(Request $request, $id)
-    {
-        $review = Review::findOrFail($id);
-        $review->update($request->all());
-        return response()->json($review);
-    }
+        if (!$user) {
+            return response()->json(['message' => 'Utilisateur non trouvé.'], 404);
+        }
 
-    // Méthode pour supprimer un avis
-    public function destroy($id)
-    {
-        $review = Review::findOrFail($id);
+        $reviews = Review::where('reviewed_user_id', $userId)->get();
+
+        return response()->json($reviews);
+    }
+    // suppresion d'avis 
+    public function destroy(Review $review)
+{
+    // Vérifiez si l'utilisateur actuel est l'auteur de l'avis
+    if (auth()->user()->id === $review->user_id) {
         $review->delete();
-        return response()->json(null, 204);
+        return response()->json(['message' => 'Avis supprimé avec succès'], 200);
+    } else {
+        return response()->json(['message' => 'Vous n\'êtes pas autorisé à supprimer cet avis'], 403);
     }
-
-
 }
 
+}
